[FIXED] InventoryView not loading bug
